containers/gam-manager/requirements.txt
Flask==2.1.
requests==2.28.1
requests-toolbelt==0.9.1
gunicorn==20.1.0
Werkzeug==2.1.2
pycdlib==1.10.0
python-magic==0.4.27
